size: make a best effort to fulfill size request
Under certain circumstances we cannot accommodate the size request:  the
specified size is either greater than or exactly equal to the available
length.  This is very likely to happen if the user is specifying less
precise size values such as MiB, where the number argument is multiplied
by a factor or three.  Limit this to the actual partition length.

It also happens that if you specify the precise number of bytes free, you
also, you'll hit an assertion in ped_geometry_test_sector_inside() so
we'll bump the request down by one sector and run with that.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>